minor cleanup + reorganization
