+++ megatron/model/__init__.py
@@ -1,6 +1,6 @@
 # coding=utf-8
-#
-# Copyright 2021 Biderman et al. This file is based on code by the authors denoted below and has been modified from its original version.
+# Copyright (c) 2021, EleutherAI contributors
+# This file is based on code by the authors denoted below and has been modified from its original version.
 #
 # Copyright (c) 2020, NVIDIA CORPORATION.  All rights reserved.
 #
@@ -19,4 +19,3 @@
 from .gpt2_model import GPT2Model, GPT2ModelPipe
 from .utils import get_params_for_weight_decay_optimization
 from .language_model import get_language_model
-from .norms import RMSNorm, ScaleNorm, LayerNorm
