+++ megatron/model/transformer.py
@@ -26,11 +26,13 @@ from .norms import LayerNorm, RMSNorm, ScaleNorm
 from megatron import get_args
 from megatron import mpu
 from megatron.module import MegatronModule
+from megatron.model.fused_bias_dropout import bias_dropout_add_fused_train, bias_dropout_add_fused_inference, \
+    get_bias_dropout_add
 from megatron.model.fused_softmax import FusedScaleMaskSoftmax
 from megatron.model.fused_bias_gelu import bias_gelu_impl
 from megatron.model.utils import openai_gelu, erf_gelu, exists
 from megatron.mpu import ParallelRelativePositionBias
-from megatron.model.positional_embeddings import SinusoidalPositionalEmbedding, RotaryEmbedding, apply_rotary_pos_emb
+from megatron.model.positional_embeddings import RotaryEmbedding, apply_rotary_pos_emb
 
 import deepspeed
 from deepspeed.ops.sparse_attention import SparseSelfAttention, VariableSparsityConfig
@@ -40,6 +42,7 @@ torch._C._jit_set_profiling_mode(False)
 torch._C._jit_set_profiling_executor(False)
 torch._C._jit_override_can_fuse_on_cpu(True)
 torch._C._jit_override_can_fuse_on_gpu(True)
+
 
 """ We use the following notation throughout this file:
      h: hidden size
@@ -62,100 +65,8 @@ torch._C._jit_override_can_fuse_on_gpu(True)
                                      unmaksed-attention-scores, attention-mask)
 """
 
-class GEGLU(MegatronModule):
 
-    def __init__(self):
-        super(GEGLU, self).__init__()
-        args = get_args()
-        self.bias_gelu_fusion = args.bias_gelu_fusion
-        self.activation_func = F.gelu
-        if args.openai_gelu:
-            self.activation_func = openai_gelu
-        elif args.onnx_safe:
-            self.activation_func = erf_gelu
-
-    def forward(self, x, bias=None):
-        x, gate = x.chunk(2, dim=-1)
-        if bias is not None:
-            bias_1, bias_2 = bias.chunk(2, dim=-1)
-            x = x + bias_1
-        else:
-            bias_1 = bias_2 = 0
-        if self.bias_gelu_fusion:
-            intermediate_parallel = \
-                bias_gelu_impl(gate, bias_2)
-        else:
-            intermediate_parallel = \
-                self.activation_func(gate + bias_2)
-        return intermediate_parallel * x
-
-
-class ParallelMLP(MegatronModule):
-    """MLP.
-
-    MLP will take the input with h hidden state, project it to 4*h
-    hidden dimension, perform nonlinear transformation, and project the
-    state back into h hidden dimension. At the end, dropout is also
-    applied.
-    """
-
-    def __init__(self, init_method, output_layer_init_method):
-        super(ParallelMLP, self).__init__()
-        args = get_args()
-
-        if args.geglu:
-            self.activation_type = "geglu"
-            mult = 8
-            self.activation_func = GEGLU()
-        else:
-            self.activation_type = "gelu"
-            mult = 4
-            self.bias_gelu_fusion = args.bias_gelu_fusion
-            self.activation_func = F.gelu
-            if args.openai_gelu:
-                self.activation_func = openai_gelu
-            elif args.onnx_safe:
-                self.activation_func = erf_gelu
-
-        # Project to 4h.
-        self.dense_h_to_4h = mpu.ColumnParallelLinear(
-            args.hidden_size,
-            mult * args.hidden_size,
-            gather_output=False,
-            init_method=init_method,
-            skip_bias_add=True)
-
-        # Project back to h.
-        self.dense_4h_to_h = mpu.RowParallelLinear(
-            4 * args.hidden_size,
-            args.hidden_size,
-            input_is_parallel=True,
-            init_method=output_layer_init_method,
-            skip_bias_add=True)
-
-    def forward(self, hidden_states):
-
-        # [s, b, 4hp]
-        intermediate_parallel, bias_parallel = self.dense_h_to_4h(hidden_states)
-
-        if self.activation_type == "gelu":
-            if self.bias_gelu_fusion:
-                intermediate_parallel = \
-                    bias_gelu_impl(intermediate_parallel, bias_parallel)
-            else:
-                intermediate_parallel = \
-                    self.activation_func(intermediate_parallel + bias_parallel)
-        elif self.activation_type == "geglu":
-            intermediate_parallel = \
-                self.activation_func(intermediate_parallel)
-        else:
-            raise ValueError(f'Activation type {self.activation_type} not recognized')
-
-        # [s, b, h]
-        output, output_bias = self.dense_4h_to_h(intermediate_parallel)
-        return output, output_bias
-
-
+# Self Attention + Transformer blocks
 class ParallelSelfAttention(MegatronModule):
     """Parallel self-attention layer abstract class.
 
@@ -289,7 +200,7 @@ class ParallelSelfAttention(MegatronModule):
                 query_rot, key_rot = query_layer, key_layer
 
             query_layer, key_layer = apply_rotary_pos_emb(query_rot, key_rot, cos, sin)
-           
+
             if exists(self.rotary_ndims):
                 query_layer = torch.cat((query_layer, query_pass), dim=-1)
                 key_layer = torch.cat((key_layer, key_pass), dim=-1)
@@ -434,32 +345,6 @@ class ParallelSelfAttention(MegatronModule):
         return output, bias
 
 
-def bias_dropout_add(x, bias, residual, prob, training):
-    # type: (Tensor, Tensor, Tensor, float, bool) -> Tensor
-    out = torch.nn.functional.dropout(x + bias, p=prob, training=training)
-    out = residual + out
-    return out
-
-
-def get_bias_dropout_add(training):
-    def _bias_dropout_add(x, bias, residual, prob):
-        return bias_dropout_add(x, bias, residual, prob, training)
-
-    return _bias_dropout_add
-
-
-@torch.jit.script
-def bias_dropout_add_fused_train(x, bias, residual, prob):
-    # type: (Tensor, Tensor, Tensor, float) -> Tensor
-    return bias_dropout_add(x, bias, residual, prob, True)
-
-
-@torch.jit.script
-def bias_dropout_add_fused_inference(x, bias, residual, prob):
-    # type: (Tensor, Tensor, Tensor, float) -> Tensor
-    return bias_dropout_add(x, bias, residual, prob, False)
-
-
 class ParallelTransformerLayer(MegatronModule):
     """A single transformer layer.
 
@@ -502,7 +387,7 @@ class ParallelTransformerLayer(MegatronModule):
                                                rpe=rpe,
                                                get_key_value=self.get_key_value,
                                                rotary=rotary)
-          
+
         self.hidden_dropout = args.hidden_dropout
         self.bias_dropout_fusion = args.bias_dropout_fusion
 
@@ -580,7 +465,7 @@ class ParallelTransformerLayer(MegatronModule):
 
         return output
 
-      
+
 class ParallelTransformer(MegatronModule):
     """Transformer class."""
 
@@ -605,8 +490,8 @@ class ParallelTransformer(MegatronModule):
 
         if args.pos_emb == 'rpe':
             rpe_emb = ParallelRelativePositionBias(causal=True, num_buckets=args.rpe_num_buckets,
-                                                        max_distance=args.rpe_max_distance,
-                                                        heads=args.num_attention_heads)
+                                                   max_distance=args.rpe_max_distance,
+                                                   heads=args.num_attention_heads)
 
         # Transformer layers.
         sparsity = args.sparsity
@@ -622,8 +507,8 @@ class ParallelTransformer(MegatronModule):
                 raise ValueError(f'Sparsity type {sparsity} not recognized')
             return ParallelTransformerLayer(
                 attention_mask_func, init_method,
-                output_layer_init_method, layer_number, sparse=sparse, 
-                rpe=rpe_emb if args.pos_emb == 'rpe' else None, 
+                output_layer_init_method, layer_number, sparse=sparse,
+                rpe=rpe_emb if args.pos_emb == 'rpe' else None,
                 get_key_value=get_key_value,
                 rotary=args.pos_emb == 'rotary')
 
@@ -693,8 +578,7 @@ class ParallelTransformer(MegatronModule):
 
         return hidden_states
 
-
-    def forward(self, hidden_states, attention_mask, layer_past=None,):
+    def forward(self, hidden_states, attention_mask, layer_past=None, ):
         # Checks
         if layer_past is not None and layer_past.numel() > 0:
             assert self.get_key_value, \
@@ -739,13 +623,111 @@ class ParallelTransformer(MegatronModule):
 
         return output
 
+
+# MLP & Activation Fns
+class GEGLU(MegatronModule):
+
+    def __init__(self):
+        super(GEGLU, self).__init__()
+        args = get_args()
+        self.bias_gelu_fusion = args.bias_gelu_fusion
+        self.activation_func = F.gelu
+        if args.openai_gelu:
+            self.activation_func = openai_gelu
+        elif args.onnx_safe:
+            self.activation_func = erf_gelu
+
+    def forward(self, x, bias=None):
+        x, gate = x.chunk(2, dim=-1)
+        if bias is not None:
+            bias_1, bias_2 = bias.chunk(2, dim=-1)
+            x = x + bias_1
+        else:
+            bias_1 = bias_2 = 0
+        if self.bias_gelu_fusion:
+            intermediate_parallel = \
+                bias_gelu_impl(gate, bias_2)
+        else:
+            intermediate_parallel = \
+                self.activation_func(gate + bias_2)
+        return intermediate_parallel * x
+
+
+class ParallelMLP(MegatronModule):
+    """MLP.
+
+    MLP will take the input with h hidden state, project it to 4*h
+    hidden dimension, perform nonlinear transformation, and project the
+    state back into h hidden dimension. At the end, dropout is also
+    applied.
+    """
+
+    def __init__(self, init_method, output_layer_init_method):
+        super(ParallelMLP, self).__init__()
+        args = get_args()
+
+        if args.geglu:
+            self.activation_type = "geglu"
+            mult = 8
+            self.activation_func = GEGLU()
+        else:
+            self.activation_type = "gelu"
+            mult = 4
+            self.bias_gelu_fusion = args.bias_gelu_fusion
+            self.activation_func = F.gelu
+            if args.openai_gelu:
+                self.activation_func = openai_gelu
+            elif args.onnx_safe:
+                self.activation_func = erf_gelu
+
+        # Project to 4h.
+        self.dense_h_to_4h = mpu.ColumnParallelLinear(
+            args.hidden_size,
+            mult * args.hidden_size,
+            gather_output=False,
+            init_method=init_method,
+            skip_bias_add=True)
+
+        # Project back to h.
+        self.dense_4h_to_h = mpu.RowParallelLinear(
+            4 * args.hidden_size,
+            args.hidden_size,
+            input_is_parallel=True,
+            init_method=output_layer_init_method,
+            skip_bias_add=True)
+
+    def forward(self, hidden_states):
+
+        # [s, b, 4hp]
+        intermediate_parallel, bias_parallel = self.dense_h_to_4h(hidden_states)
+
+        if self.activation_type == "gelu":
+            if self.bias_gelu_fusion:
+                intermediate_parallel = \
+                    bias_gelu_impl(intermediate_parallel, bias_parallel)
+            else:
+                intermediate_parallel = \
+                    self.activation_func(intermediate_parallel + bias_parallel)
+        elif self.activation_type == "geglu":
+            intermediate_parallel = \
+                self.activation_func(intermediate_parallel)
+        else:
+            raise ValueError(f'Activation type {self.activation_type} not recognized')
+
+        # [s, b, h]
+        output, output_bias = self.dense_4h_to_h(intermediate_parallel)
+        return output, output_bias
+
+
+# Pipeline Parallel Modules
 class ParallelTransformerLayerPipe(ParallelTransformerLayer):
     """Extends ParallelTransformerLayer to forward attention_mask through the pipeline. """
 
     def forward(self, args):
-        in_inference = len(args) in [4,5] # length of the args in inference can either be 4 (no rotary pos emb) or 5
-        in_train = len(args) in [2,3] # length of the args in training can either be 2 (no rotary pos emb) or 3
-        has_rotary_pos_emb = len(args) in [3, 5] # if we're passing around a rotary pos emb, length of args will either be 3 or 5
+        in_inference = len(args) in [4, 5]  # length of the args in inference can either be 4 (no rotary pos emb) or 5
+        in_train = len(args) in [2, 3]  # length of the args in training can either be 2 (no rotary pos emb) or 3
+        has_rotary_pos_emb = len(args) in [3,
+                                           5]  # if we're passing around a rotary pos emb, length of args will either be 3 or 5
 
         if in_train:
             hidden_states, attention_mask = args
@@ -770,7 +752,8 @@ class ParallelTransformerLayerPipe(ParallelTransformerLayer):
                 hidden_states = outputs
             return hidden_states, layer_past, presents, attention_mask
         else:
-            raise ValueError(f'In layer {self.layer_number} - Incorrect number of arguments ({len(args)}) for {self.__class__.__name__}')
+            raise ValueError(
+                f'In layer {self.layer_number} - Incorrect number of arguments ({len(args)}) for {self.__class__.__name__}')
 
 
 class NormPipe(MegatronModule):
@@ -793,162 +776,7 @@ class NormPipe(MegatronModule):
         else:
             raise ValueError(f'Incorrect number of arguments for {self.__class__.__name__}')
 
-class Embedding(MegatronModule):
-    """Language model embeddings.
-    Arguments:
-        hidden_size: hidden size
-        vocab_size: vocabulary size
-        max_sequence_length: maximum size of sequence. This
-                             is used for positional embedding
-        embedding_dropout_prob: dropout probability for embeddings
-        init_method: weight initialization method
-        num_tokentypes: size of the token-type embeddings. 0 value
-                        will ignore this embedding
-    """
 
-    def __init__(self,
-                 hidden_size,
-                 vocab_size,
-                 max_sequence_length,
-                 embedding_dropout_prob,
-                 init_method,
-                 num_tokentypes=0):
-        super(Embedding, self).__init__()
-        args = get_args()
-        self.hidden_size = hidden_size
-        self.init_method = init_method
-        self.num_tokentypes = num_tokentypes
-
-        # Word embeddings (parallel).
-        self.word_embeddings = mpu.VocabParallelEmbedding(
-            vocab_size, self.hidden_size, init_method=self.init_method)
-        self._word_embeddings_key = 'word_embeddings'
-
-        # Position embedding (serial).
-        self.embedding_type = args.pos_emb
-        if self.embedding_type == "learned":
-            self.position_embeddings = torch.nn.Embedding(
-                max_sequence_length, self.hidden_size)
-            self._position_embeddings_key = 'position_embeddings'
-            # Initialize the position embeddings.
-            self.init_method(self.position_embeddings.weight)
-        elif self.embedding_type == "sinusoidal":
-            self.position_embeddings = SinusoidalPositionalEmbedding(self.hidden_size)
-
-        # Token type embedding.
-        # Add this as an optional field that can be added through
-        # method call so we can load a pretrain model without
-        # token types and add them as needed.
-        self._tokentype_embeddings_key = 'tokentype_embeddings'
-        if self.num_tokentypes > 0:
-            self.tokentype_embeddings = torch.nn.Embedding(self.num_tokentypes,
-                                                           self.hidden_size)
-            # Initialize the token-type embeddings.
-            self.init_method(self.tokentype_embeddings.weight)
-        else:
-            self.tokentype_embeddings = None
-
-        # Embeddings dropout
-        self.embedding_dropout = torch.nn.Dropout(embedding_dropout_prob)
-
-    def add_tokentype_embeddings(self, num_tokentypes):
-        """Add token-type embedding. This function is provided so we can add
-        token-type embeddings in case the pretrained model does not have it.
-        This allows us to load the model normally and then add this embedding.
-        """
-        if self.tokentype_embeddings is not None:
-            raise Exception('tokentype embeddings is already initialized')
-        if torch.distributed.get_rank() == 0:
-            print('adding embedding for {} tokentypes'.format(num_tokentypes),
-                  flush=True)
-        self.num_tokentypes = num_tokentypes
-        self.tokentype_embeddings = torch.nn.Embedding(num_tokentypes,
-                                                       self.hidden_size)
-        # Initialize the token-type embeddings.
-        self.init_method(self.tokentype_embeddings.weight)
-
-    def forward(self, input_ids, position_ids, tokentype_ids=None):
-        # Embeddings.
-        words_embeddings = self.word_embeddings(input_ids)
-        if self.embedding_type in ["learned", "sinusoidal"]:
-            position_embeddings = self.position_embeddings(position_ids)
-            embeddings = words_embeddings + position_embeddings
-        else:
-            embeddings = words_embeddings
-        if tokentype_ids is not None:
-            assert self.tokentype_embeddings is not None
-            embeddings = embeddings + self.tokentype_embeddings(tokentype_ids)
-        else:
-            assert self.tokentype_embeddings is None
-
-        # Dropout.
-        embeddings = self.embedding_dropout(embeddings)
-        return embeddings
-
-    def state_dict_for_save_checkpoint(self, destination=None, prefix='',
-                                       keep_vars=False):
-        """For easy load."""
-
-        state_dict_ = {}
-        state_dict_[self._word_embeddings_key] \
-            = self.word_embeddings.state_dict(destination, prefix, keep_vars)
-        if self.embedding_type == "learned":
-            state_dict_[self._position_embeddings_key] \
-                = self.position_embeddings.state_dict(
-                destination, prefix, keep_vars)
-        if self.num_tokentypes > 0:
-            state_dict_[self._tokentype_embeddings_key] \
-                = self.tokentype_embeddings.state_dict(
-                destination, prefix, keep_vars)
-
-        return state_dict_
-
-    def load_state_dict(self, state_dict, strict=True):
-        """Customized load."""
-
-        # Word embedding.
-        if self._word_embeddings_key in state_dict:
-            state_dict_ = state_dict[self._word_embeddings_key]
-        else:
-            # for backward compatibility.
-            state_dict_ = {}
-            for key in state_dict.keys():
-                if 'word_embeddings' in key:
-                    state_dict_[key.split('word_embeddings.')[1]] \
-                        = state_dict[key]
-        self.word_embeddings.load_state_dict(state_dict_, strict=strict)
-
-        # Position embedding.
-        if self.embedding_type == "learned":
-            if self._position_embeddings_key in state_dict:
-                state_dict_ = state_dict[self._position_embeddings_key]
-            else:
-                # for backward compatibility.
-                state_dict_ = {}
-                for key in state_dict.keys():
-                    if 'position_embeddings' in key:
-                        state_dict_[key.split('position_embeddings.')[1]] \
-                            = state_dict[key]
-            self.position_embeddings.load_state_dict(state_dict_, strict=strict)
-
-        # Tokentype embedding.
-        if self.num_tokentypes > 0:
-            state_dict_ = {}
-            if self._tokentype_embeddings_key in state_dict:
-                state_dict_ = state_dict[self._tokentype_embeddings_key]
-            else:
-                # for backward compatibility.
-                for key in state_dict.keys():
-                    if 'tokentype_embeddings' in key:
-                        state_dict_[key.split('tokentype_embeddings.')[1]] \
-                            = state_dict[key]
-            if len(state_dict_.keys()) > 0:
-                self.tokentype_embeddings.load_state_dict(state_dict_,
-                                                          strict=strict)
-            else:
-                print('***WARNING*** expected tokentype embeddings in the '
-                      'checkpoint but could not find it', flush=True)
-                        
 class RowParallelLinearPipe(mpu.RowParallelLinear):
     """Another helper class to pass presents through to the output when doing inference with a Pipe Parallel model"""
 
@@ -965,33 +793,3 @@ class RowParallelLinearPipe(mpu.RowParallelLinear):
             return logits, presents
         else:
             raise ValueError(f'Incorrect number of arguments for {self.__class__.__name__}')
-
-
-class EmbeddingPipe(Embedding):
-    """Extends Embedding to forward attention_mask through the pipeline."""
-
-    @property
-    def word_embeddings_weight(self):
-        """Easy accessory for the pipeline engine to tie embeddings across stages."""
-        return self.word_embeddings.weight
-
-    def forward(self, args):
-        in_inference = len(args) == 4 # if the length of the args is 4, we're in inference :|
-        in_train = len(args) == 3
-
-        input_ids = args[0]
-        position_ids = args[1]
-        attention_mask = args[2]
-        if in_inference:
-            layer_past = args[3]
-        elif in_train:
-            pass
-        else:
-            raise ValueError(f'Incorrect number of args passed to {self.__class__.__name__}')
-
-        embeddings = super().forward(input_ids, position_ids)
-        if in_inference:
-            return embeddings, layer_past, attention_mask
-        else:
-            return embeddings, attention_mask
-
